Add files via upload
This is final project in the Data Analysis Nanodegree program at Udacity.  Tableau was used to make an explanatory analysis of the data given by creating visualizations.  I believe there is always more that can be done with a data and should you have suggestions, please do not hesitate to forward them. As I am always open to new perspectives to better understand the given data and the information that can be extracted form it!